fix store author

+++ src/main/java/com/peak/main/controller/store.java
@@ -1,13 +1,16 @@
 package com.peak.main.controller;
 
+import com.peak.Util.Role;
+import com.peak.main.Request.RequestItem;
 import com.peak.main.Request.Response;
 import com.peak.main.model.Store;
-import com.peak.main.repository.StoreRepository;
+import com.peak.main.model.User;
 import com.peak.main.service.StoreService;
 import lombok.AllArgsConstructor;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -32,28 +35,80 @@ public class store {
 /*  /store
     {
         "name":"name",
-        "userID":"1",
         "detail":"detail",
         "image":"http"
     }
  */
-    @PostMapping
+    @PostMapping("/me")
     @PreAuthorize("hasAnyRole('ADMIN', 'SELLER')")
-    public ResponseEntity<Response> add(@RequestBody Store store) {
+    public ResponseEntity<Response> add(@RequestBody Store store, Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        if (storeService.findByUserId(user.getId()).isPresent()) return ResponseEntity.ok(new Response("Already exists"));
         if (store.getName() == null ||
-                store.getUserID() == null ||
                 store.getDetail() == null ||
                 store.getImage() == null)
             return ResponseEntity.notFound().build();
+        store.setUserID(user.getId());
 
         return ResponseEntity.ok(new Response(storeService.save(store)));
     }
 
 //  /store/{id}
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasAnyRole('ADMIN')")
-    public ResponseEntity<Response> delete(@PathVariable Long id) {
-        storeService.deleteById(id);
-        return ResponseEntity.ok().build();
+    @PreAuthorize("hasAnyRole('ADMIN','SELLER')")
+    public ResponseEntity<Response> delete(@PathVariable Long id, Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        if (user.getRole().equals(Role.ADMIN) || storeService.hasPermitionStore(user.getId(), id)) {
+            storeService.deleteById(id);
+            return ResponseEntity.ok(new Response("[]"));
+        }
+
+        return ResponseEntity.status(403).build();
     }
+
+/*  /store/{id}
+    {
+        "name":"name",
+        "cost":"12",
+        "category":"big",
+        "detail":"detail",
+        "stock":"1",            // could be null
+        "sold":"1",             // could be null
+        "discount":"1"          // could be null
+        "types":["red","blue"],  // could be null
+        "images":["image1"]     // could be null
+    }
+ */
+    @PostMapping("/{id}")
+    @PreAuthorize("hasAnyRole('ADMIN','SELLER')")
+    public ResponseEntity<Response> add(@PathVariable Long id, @RequestBody RequestItem item, Authentication authentication) {
+        if (item.getName() == null ||
+                item.getCost() == null ||
+                item.getCategory() == null ||
+                item.getDetail() == null)
+            return ResponseEntity.notFound().build();
+
+        item.setStoreID(id);
+
+        User user = (User) authentication.getPrincipal();
+        if (user.getRole().equals(Role.ADMIN)) return ResponseEntity.ok(new Response(storeService.saveToStore(item)));
+        if (storeService.hasPermitionStore(user.getId(), id)) return ResponseEntity.ok(new Response(storeService.saveToStore(item)));
+
+        return ResponseEntity.status(403).build();
+    }
+
+//  /store/{storeId}/item/{itemId}
+    @DeleteMapping("/{sid}/item/{iid}")
+    @PreAuthorize("hasAnyRole('ADMIN','SELLER')")
+    public ResponseEntity<Response> delete(@PathVariable Long iid, @PathVariable Long sid, Authentication authentication) {
+        User user = (User) authentication.getPrincipal();
+        if (!storeService.hasPermitionItem(sid, iid)) return ResponseEntity.status(403).build();
+
+        if (user.getRole().equals(Role.ADMIN) || storeService.hasPermitionStore(user.getId(), sid)) {
+            storeService.deleteFromStore(iid);
+            return ResponseEntity.ok(new Response("[]"));
+        }
+        return ResponseEntity.status(403).build();
+    }
+
 }
